+++ execution_engine/src/core/engine_state/execution_result.rs
@@ -5,8 +5,8 @@ use casper_types::{bytesrepr::FromBytes, CLTyped, CLValue, CLValueError, Key, Tr
 use super::{error, execution_effect::ExecutionEffect, op::Op};
 use crate::{
     shared::{
-        additive_map::AdditiveMap, gas::Gas, motes::Motes, newtypes::CorrelationId,
-        stored_value::StoredValue, transform::Transform,
+        additive_map::AdditiveMap, execution_journal::ExecutionJournal, gas::Gas, motes::Motes,
+        newtypes::CorrelationId, stored_value::StoredValue, transform::Transform,
     },
     storage::global_state::StateReader,
 };
@@ -16,31 +16,21 @@ fn make_payment_error_effects(
     account_main_purse_balance: Motes,
     account_main_purse_balance_key: Key,
     proposer_main_purse_balance_key: Key,
-) -> Result<ExecutionEffect, CLValueError> {
-    let mut ops = AdditiveMap::new();
-    let mut transforms = AdditiveMap::new();
-
+) -> Result<ExecutionJournal, CLValueError> {
     let new_balance = account_main_purse_balance - max_payment_cost;
     // from_t for U512 is assumed to never panic
-    let new_balance_clvalue = CLValue::from_t(new_balance.value())?;
-    let new_balance_value = StoredValue::CLValue(new_balance_clvalue);
-
-    let account_main_purse_balance_normalize = account_main_purse_balance_key.normalize();
-    let proposer_main_purse_balance_normalize = proposer_main_purse_balance_key.normalize();
-
-    ops.insert(account_main_purse_balance_normalize, Op::Write);
-    transforms.insert(
-        account_main_purse_balance_normalize,
-        Transform::Write(new_balance_value),
-    );
-
-    ops.insert(proposer_main_purse_balance_normalize, Op::Add);
-    transforms.insert(
-        proposer_main_purse_balance_normalize,
-        Transform::AddUInt512(max_payment_cost.value()),
-    );
-
-    Ok(ExecutionEffect::new(ops, transforms))
+    let new_balance_value = StoredValue::CLValue(CLValue::from_t(new_balance.value())?);
+    Ok(vec![
+        (
+            account_main_purse_balance_key.normalize(),
+            Transform::Write(new_balance_value),
+        ),
+        (
+            proposer_main_purse_balance_key.normalize(),
+            Transform::AddUInt512(max_payment_cost.value()),
+        ),
+    ]
+    .into())
 }
 
 #[derive(Clone, Debug)]
@@ -51,21 +41,24 @@ pub enum ExecutionResult {
         execution_effect: ExecutionEffect,
         transfers: Vec<TransferAddr>,
         cost: Gas,
+        execution_journal: ExecutionJournal,
     },
     /// Execution was finished successfully
     Success {
         execution_effect: ExecutionEffect,
         transfers: Vec<TransferAddr>,
         cost: Gas,
+        execution_journal: ExecutionJournal,
     },
 }
 
 impl Default for ExecutionResult {
     fn default() -> Self {
         ExecutionResult::Success {
-            execution_effect: ExecutionEffect::default(),
-            transfers: Vec::default(),
-            cost: Gas::default(),
+            execution_journal: Default::default(),
+            execution_effect: Default::default(),
+            transfers: Default::default(),
+            cost: Default::default(),
         }
     }
 }
@@ -92,6 +85,7 @@ impl ExecutionResult {
             execution_effect: Default::default(),
             transfers: Vec::default(),
             cost: Gas::default(),
+            execution_journal: Default::default(),
         }
     }
 
@@ -151,21 +145,25 @@ impl ExecutionResult {
                 error,
                 execution_effect,
                 transfers,
+                execution_journal,
                 ..
             } => ExecutionResult::Failure {
                 error,
                 execution_effect,
                 transfers,
                 cost,
+                execution_journal,
             },
             ExecutionResult::Success {
                 execution_effect,
                 transfers,
+                execution_journal,
                 ..
             } => ExecutionResult::Success {
                 execution_effect,
                 transfers,
                 cost,
+                execution_journal,
             },
         }
     }
@@ -176,19 +174,25 @@ impl ExecutionResult {
                 error,
                 cost,
                 transfers,
+                execution_journal,
                 ..
             } => ExecutionResult::Failure {
                 error,
                 execution_effect,
                 transfers,
                 cost,
+                execution_journal,
             },
             ExecutionResult::Success {
-                cost, transfers, ..
+                cost,
+                transfers,
+                execution_journal,
+                ..
             } => ExecutionResult::Success {
                 execution_effect,
                 transfers,
                 cost,
+                execution_journal,
             },
         }
     }
@@ -199,21 +203,25 @@ impl ExecutionResult {
                 error,
                 execution_effect,
                 cost,
+                execution_journal,
                 ..
             } => ExecutionResult::Failure {
                 error,
                 execution_effect,
                 transfers,
                 cost,
+                execution_journal,
             },
             ExecutionResult::Success {
                 cost,
                 execution_effect,
+                execution_journal,
                 ..
             } => ExecutionResult::Success {
                 execution_effect,
                 transfers,
                 cost,
+                execution_journal,
             },
         }
     }
@@ -271,16 +279,18 @@ impl ExecutionResult {
         account_main_purse_balance_key: Key,
         proposer_main_purse_balance_key: Key,
     ) -> Result<ExecutionResult, CLValueError> {
-        let execution_effect = make_payment_error_effects(
+        let execution_journal = make_payment_error_effects(
             max_payment_cost,
             account_main_purse_balance,
             account_main_purse_balance_key,
             proposer_main_purse_balance_key,
         )?;
+        let execution_effect: ExecutionEffect = execution_journal.clone().into();
         let transfers = Vec::default();
         Ok(ExecutionResult::Failure {
             error,
             execution_effect,
+            execution_journal,
             transfers,
             cost: gas_cost,
         })
@@ -295,26 +305,28 @@ impl ExecutionResult {
     }
 }
 
-impl From<&ExecutionResult> for casper_types::ExecutionResult {
-    fn from(ee_execution_result: &ExecutionResult) -> Self {
+impl From<ExecutionResult> for casper_types::JsonExecutionResult {
+    fn from(ee_execution_result: ExecutionResult) -> Self {
         match ee_execution_result {
             ExecutionResult::Success {
-                execution_effect,
+                execution_effect: _,
                 transfers,
                 cost,
-            } => casper_types::ExecutionResult::Success {
-                effect: execution_effect.into(),
-                transfers: transfers.clone(),
+                execution_journal,
+            } => casper_types::JsonExecutionResult::Success {
+                effect: execution_journal.into(),
+                transfers,
                 cost: cost.value(),
             },
             ExecutionResult::Failure {
                 error,
-                execution_effect,
+                execution_effect: _,
                 transfers,
                 cost,
-            } => casper_types::ExecutionResult::Failure {
-                effect: execution_effect.into(),
-                transfers: transfers.clone(),
+                execution_journal,
+            } => casper_types::JsonExecutionResult::Failure {
+                effect: execution_journal.into(),
+                transfers,
                 cost: cost.value(),
                 error_message: error.to_string(),
             },
@@ -405,6 +417,7 @@ impl ExecutionResultBuilder {
 
         let mut ret: ExecutionResult = ExecutionResult::Success {
             execution_effect: Default::default(),
+            execution_journal: Default::default(),
             transfers,
             cost,
         };