+++ execution_engine/src/shared.rs
@@ -1,8 +1,11 @@
 #![allow(missing_docs)]
 
+pub use type_mismatch::TypeMismatch;
+
 pub mod additive_map;
 #[macro_use]
 pub mod gas;
+pub mod execution_journal;
 pub mod host_function_costs;
 pub mod logging;
 pub mod motes;
@@ -19,5 +22,3 @@ pub mod utils;
 pub mod wasm;
 pub mod wasm_config;
 pub mod wasm_prep;
-
-pub use type_mismatch::TypeMismatch;