+++ types/src/lib.rs
@@ -21,6 +21,49 @@ extern crate alloc;
 #[macro_use]
 extern crate std;
 
+pub use access_rights::{AccessRights, ACCESS_RIGHTS_SERIALIZED_LENGTH};
+#[doc(inline)]
+pub use api_error::ApiError;
+pub use block_time::{BlockTime, BLOCKTIME_SERIALIZED_LENGTH};
+pub use cl_type::{named_key_type, CLType, CLTyped};
+pub use cl_value::{CLTypeMismatch, CLValue, CLValueError};
+pub use contract_wasm::{ContractWasm, ContractWasmHash};
+#[doc(inline)]
+pub use contracts::{
+    Contract, ContractHash, ContractPackage, ContractPackageHash, ContractVersion,
+    ContractVersionKey, EntryPoint, EntryPointAccess, EntryPointType, EntryPoints, Group,
+    Parameter,
+};
+pub use crypto::*;
+pub use deploy_info::DeployInfo;
+pub use execution_result::{
+    JsonExecutionJournal, JsonExecutionResult, OpKind, Operation, Transform, TransformEntry,
+};
+pub use json_pretty_printer::json_pretty_print;
+#[doc(inline)]
+pub use key::{
+    DictionaryAddr, HashAddr, Key, KeyTag, BLAKE2B_DIGEST_LENGTH, DICTIONARY_ITEM_KEY_MAX_LENGTH,
+    KEY_DICTIONARY_LENGTH, KEY_HASH_LENGTH,
+};
+pub use named_key::NamedKey;
+pub use phase::{Phase, PHASE_SERIALIZED_LENGTH};
+pub use protocol_version::{ProtocolVersion, VersionCheckResult};
+#[doc(inline)]
+pub use runtime_args::{NamedArg, RuntimeArgs};
+pub use semver::{ParseSemVerError, SemVer, SEM_VER_SERIALIZED_LENGTH};
+pub use tagged::Tagged;
+pub use transfer::{
+    DeployHash, FromStrError as TransferFromStrError, Transfer, TransferAddr, DEPLOY_HASH_LENGTH,
+    TRANSFER_ADDR_LENGTH,
+};
+pub use transfer_result::{TransferResult, TransferredTo};
+pub use uref::{FromStrError as URefFromStrError, URef, UREF_ADDR_LENGTH, UREF_SERIALIZED_LENGTH};
+
+pub use crate::{
+    era_id::EraId,
+    uint::{UIntParseError, U128, U256, U512},
+};
+
 mod access_rights;
 pub mod account;
 pub mod api_error;
@@ -49,49 +92,6 @@ mod transfer;
 mod transfer_result;
 mod uint;
 mod uref;
-
-pub use access_rights::{AccessRights, ACCESS_RIGHTS_SERIALIZED_LENGTH};
-#[doc(inline)]
-pub use api_error::ApiError;
-pub use block_time::{BlockTime, BLOCKTIME_SERIALIZED_LENGTH};
-pub use cl_type::{named_key_type, CLType, CLTyped};
-pub use cl_value::{CLTypeMismatch, CLValue, CLValueError};
-pub use contract_wasm::{ContractWasm, ContractWasmHash};
-#[doc(inline)]
-pub use contracts::{
-    Contract, ContractHash, ContractPackage, ContractPackageHash, ContractVersion,
-    ContractVersionKey, EntryPoint, EntryPointAccess, EntryPointType, EntryPoints, Group,
-    Parameter,
-};
-pub use crypto::*;
-pub use deploy_info::DeployInfo;
-pub use execution_result::{
-    ExecutionEffect, ExecutionResult, OpKind, Operation, Transform, TransformEntry,
-};
-pub use json_pretty_printer::json_pretty_print;
-#[doc(inline)]
-pub use key::{
-    DictionaryAddr, HashAddr, Key, KeyTag, BLAKE2B_DIGEST_LENGTH, DICTIONARY_ITEM_KEY_MAX_LENGTH,
-    KEY_DICTIONARY_LENGTH, KEY_HASH_LENGTH,
-};
-pub use named_key::NamedKey;
-pub use phase::{Phase, PHASE_SERIALIZED_LENGTH};
-pub use protocol_version::{ProtocolVersion, VersionCheckResult};
-#[doc(inline)]
-pub use runtime_args::{NamedArg, RuntimeArgs};
-pub use semver::{ParseSemVerError, SemVer, SEM_VER_SERIALIZED_LENGTH};
-pub use tagged::Tagged;
-pub use transfer::{
-    DeployHash, FromStrError as TransferFromStrError, Transfer, TransferAddr, DEPLOY_HASH_LENGTH,
-    TRANSFER_ADDR_LENGTH,
-};
-pub use transfer_result::{TransferResult, TransferredTo};
-pub use uref::{FromStrError as URefFromStrError, URef, UREF_ADDR_LENGTH, UREF_SERIALIZED_LENGTH};
-
-pub use crate::{
-    era_id::EraId,
-    uint::{UIntParseError, U128, U256, U512},
-};
 
 #[cfg(not(any(feature = "std", feature = "no-std")))]
 compile_error!(