+++ execution_engine/CHANGELOG.md
@@ -22,6 +22,7 @@ All notable changes to this project will be documented in this file.  The format
 * Support building and testing using stable Rust.
 * Increase price of `create_purse` to 2.5CSPR.
 * Improve doc comments to clarify behavior of the bidding functionality.
+* Execution transforms are ordered by insertion order.
 
 ### Fixed
 * Fix a case where user could potentially supply a refund purse as a payment purse.