+++ node/CHANGELOG.md
@@ -26,6 +26,7 @@ All notable changes to this project will be documented in this file.  The format
   * `[gossip][gossip_request_timeout]`
   * `[gossip][get_remainder_timeout]`
   * `[fetcher][get_from_peer_timeout]`
+  * Execution transforms are ordered by insertion order.
 
 
 ### Removed